% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_test.R
\name{add_hello}
\alias{add_hello}
\title{Add together two numbers.}
\usage{
add_hello(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\examples{
add(1, 1)
add(10, 1)
}
\seealso{
\code{\link{enrichResult-class}}, \code{\link{compareCluster}}
}
\keyword{manip}

